requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.11.0=pypi_0
astroid=2.4.2=pypi_0
cached-property=1.5.2=pypi_0
cachetools=4.2.0=pypi_0
certifi=2020.12.5=py36haa95532_0
chardet=4.0.0=pypi_0
chart-studio=1.1.0=pypi_0
click=7.1.2=pypi_0
colorama=0.4.4=pypi_0
cycler=0.10.0=pypi_0
flask=1.1.2=pypi_0
google-auth=1.24.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
grpcio=1.32.0=pypi_0
gunicorn=20.0.4=pypi_0
h5py=2.10.0=pypi_0
idna=2.10=pypi_0
importlib-metadata=3.3.0=pypi_0
isort=5.7.0=pypi_0
itsdangerous=1.1.0=pypi_0
jinja2=2.11.2=pypi_0
joblib=1.0.0=pypi_0
keras=2.2.2=pypi_0
keras-applications=1.0.4=pypi_0
keras-preprocessing=1.0.2=pypi_0
kiwisolver=1.3.1=pypi_0
lazy-object-proxy=1.4.3=pypi_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.3=pypi_0
mccabe=0.6.1=pypi_0
numpy=1.18.2=pypi_0
oauthlib=3.1.0=pypi_0
opt-einsum=3.3.0=pypi_0
pandas=1.0.3=pypi_0
patsy=0.5.1=pypi_0
pillow=8.1.0=pypi_0
pip=20.3.3=py36haa95532_0
plotly=4.14.1=pypi_0
plotly-express=0.4.1=pypi_0
protobuf=3.14.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pylint=2.6.0=pypi_0
pyparsing=2.4.7=pypi_0
python=3.6.12=h5500b2f_2
python-dateutil=2.8.1=pypi_0
pytz=2020.5=pypi_0
pyyaml=5.3.1=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
retrying=1.3.3=pypi_0
rsa=4.6=pypi_0
scikit-learn=0.22=pypi_0
scipy=1.5.4=pypi_0
setuptools=51.0.0=py36haa95532_2
six=1.15.0=pypi_0
sqlite=3.33.0=h2a8f88b_0
statsmodels=0.12.1=pypi_0
tensorboard=2.4.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
toml=0.10.2=pypi_0
typed-ast=1.4.2=pypi_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.12.1=pypi_0
zipp=3.4.0=pypi_0
zlib=1.2.11=h62dcd97_4
